+++ .idea/workspace.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ChangeListManager">
+    <list default="true" id="e051ef6f-8d76-454d-bd33-387c37a437a9" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/.idea/.gitignore" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/A2_COMP346.iml" beforeDir="false" afterPath="$PROJECT_DIR$/A2_COMP346.iml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Accounts-v2.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Client-v2.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Network-v2.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/OSpa2driver-v2.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Server-v2.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Transactions-v2.java" beforeDir="false" />
+    </list>
+    <option name="SHOW_DIALOG" value="false" />
+    <option name="HIGHLIGHT_CONFLICTS" value="true" />
+    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
+    <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
+  <component name="MarkdownSettingsMigration">
+    <option name="stateVersion" value="1" />
+  </component>
+  <component name="ProjectId" id="25QRoOnOcf0on7t3Xsl1t5mtR7N" />
+  <component name="ProjectViewState">
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showLibraryContents" value="true" />
+  </component>
+  <component name="PropertiesComponent">
+    <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
+    <property name="RunOnceActivity.ShowReadmeOnStart" value="true" />
+    <property name="project.structure.last.edited" value="Problems" />
+    <property name="project.structure.proportion" value="0.15" />
+    <property name="project.structure.side.proportion" value="0.2" />
+  </component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
+  <component name="TaskManager">
+    <task active="true" id="Default" summary="Default task">
+      <changelist id="e051ef6f-8d76-454d-bd33-387c37a437a9" name="Changes" comment="" />
+      <created>1645457499849</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1645457499849</updated>
+    </task>
+    <servers />
+  </component>
+</project>
